production/requirements.txt
tqdm
more_itertools
numpy
onnxruntime-gpu
transformers
flask
requests